UPSTREAM: xskmap: Move non-standard list manipulation to helper
Add a helper in list.h for the non-standard way of clearing a list that is
used in xskmap. This makes it easier to reuse it in the other map types,
and also makes sure this usage is not forgotten in any list refactorings in
the future.

Change-Id: I3717523d2a890fad4e026aedb5066e1cc75d4d5b
Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
(cherry picked from commit 4c3219d3c02ab2df41a3c8a3824f69331f57fe0a)
Signed-off-by: z3rokwq <z3rokwq@gmail.com>