UPSTREAM: bpf: Add abnormal return checks.
LD_[ABS|IND] instructions may return from the function early. bpf_tail_call
pseudo instruction is either fallthrough or return. Allow them in the
subprograms only when subprograms are BTF annotated and have scalar return
types. Allow ld_abs and tail_call in the main program even if it calls into
subprograms. In the past that was not ok to do for ld_abs, since it was JITed
with special exit sequence. Since bpf_gen_ld_abs() was introduced the ld_abs
looks like normal exit insn from JIT point of view, so it's safe to allow them
in the main program.

Change-Id: I8ab267d2cdfd4a8fb8c99314039e7e261fd7f9f0
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Signed-off-by: z3rokwq <z3rokwq@gmail.com>